Fix 17 bugs (4 critical, 3 high, 3 medium, 7 low)
Critical:
- Title-index mismatch when clips fail (main.py) — map done paths to moments
- Resume path skipped transcript generation (main.py) — backfill from cached audio
- set_moments() overwrote transcript/crop data (pipeline_cache.py) — merge stored keys
- -hwaccel_device flag leaked during HW fallback (hwaccel.py) — skip companion flag

High:
- OpenCV VideoCapture leaked on exception in both get_crop_params_dynamic and
  _refine_transitions (cropper.py) — wrap in try/finally
- subtitle_files cleanup iterated shared list without lock (api_bridge.py)
- Unescaped commas in drawtext filter (subtitler.py) — add comma escaping

Medium:
- Mixed int/float timestamps in fallback path (detector.py) — normalize to float
- step_downloaded cache field written but never read (main.py) — remove dead code

Low:
- Hardcoded C:\\ drive for fonts (subtitler.py) — use SystemRoot env var
- Backfill audio extraction truncated to 60s (api_bridge.py) — probe clip duration
- Scheduler mutated schedule items outside lock (api_bridge.py) — wrap mutations
- reset_cancel() cleared process list without terminating (subprocess_utils.py)
- find_sentence_boundary lacked cancel checks (transcriber.py)
- generate_subtitles lacked cancel check (subtitler.py)

+++ detector.py
@@ -246,12 +246,12 @@ def _fallback_moments(
 
     clips = []
     for start in starts[:num_clips]:
-        end = min(total_seconds, int(start) + usable_duration)
+        end = min(total_seconds, float(start) + usable_duration)
         clips.append(
             {
-                "start": int(start),
-                "end": int(end),
-                "duration": int(end - start),
+                "start": float(start),
+                "end": float(end),
+                "duration": float(end - start),
                 "score": 0.0,
             }
         )
